pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "loomcli"
version = "0.3.0"
description = "Weave — the Powerloom CLI. Declarative manifests (apply/plan/destroy) + resource inspection, validated against the authoritative JSON Schema published in this repo."
readme = "README.md"
license = { text = "Proprietary" }
requires-python = ">=3.11"
authors = [
    { name = "Shane Levy", email = "shanerlevy@gmail.com" },
]
maintainers = [
    { name = "Shane Levy", email = "shanerlevy@gmail.com" },
]
keywords = [
    "powerloom",
    "weave",
    "loom",
    "cli",
    "agents",
    "claude",
    "cma",
    "infrastructure-as-code",
    "declarative",
    "manifests",
    "mcp",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: Other/Proprietary License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Systems Administration",
    "Topic :: Utilities",
]
dependencies = [
    "typer>=0.12,<1.0",
    "httpx>=0.27,<1.0",
    "pydantic>=2.9,<3.0",
    "pyyaml>=6.0,<7.0",
    "rich>=13.7,<15.0",
    "platformdirs>=4.2,<5.0",
    "jsonschema>=4.23,<5.0",
    "referencing>=0.35,<1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3,<9.0",
    "pytest-asyncio>=0.24,<1.0",
    "respx>=0.21,<1.0",
    "pyinstaller>=6.10,<7.0",
]

[project.urls]
Homepage = "https://powerloom.org"
Repository = "https://github.com/shanerlevy-debug/loomcli"
Issues = "https://github.com/shanerlevy-debug/loomcli/issues"
Documentation = "https://github.com/shanerlevy-debug/loomcli/blob/main/README.md"
Platform = "https://github.com/shanerlevy-debug/Powerloom"
Changelog = "https://github.com/shanerlevy-debug/loomcli/blob/main/CHANGELOG.md"

[project.scripts]
weave = "loomcli.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["loomcli"]

# Bundle the authoritative JSON Schema from `schema/v1/` into the wheel so
# pip-installed `weave` can validate manifests without needing the repo
# checkout. The validator's _find_schema_dir() checks this path as a
# fallback after PyInstaller bundles + in-repo path.
[tool.hatch.build.targets.wheel.force-include]
"schema/v1" = "loomcli/_bundled_schema/v1"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
